+++ src/stories/Button/Button.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import { Button } from '../../components/Button/Button';
+import { Button } from '../../components/common/Button/Button';
 
 const meta: Meta<typeof Button> = {
   title: 'Example/Button',
